requirements.txt
# Minimal setup without pandas
httpx==0.25.2
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
rich==13.7.0
pydantic>=2.5.0
sqlalchemy==2.0.23